change everything

+++ Data.Space/ScreenSize.cs
@@ -1,7 +1,7 @@
 ﻿namespace Data.Space
 {
     // This should be interpreted as "the size of screen space", i.e. the renderable area
-    public readonly record struct ScreenSize(int Width, int Height)
+    public readonly record struct ScreenSize(ushort Width, ushort Height)
     {
     }
 }
